Make Dataverse AIP mets changes
NB. The rebase should see archivematicaCreateMETS2.py revert to the
mainline version but with the Dataverse code merged into it, or a
more sustainable module created surrounding that work.
