+++ src/MCPClient/lib/clientScripts/archivematicaCreateMETS2.py
@@ -26,8 +26,10 @@ from __future__ import print_function
 import collections
 import copy
 from glob import glob
+from itertools import chain
 import lxml.etree as etree
 import os
+import pprint
 import re
 import sys
 import traceback
@@ -36,12 +38,11 @@ from uuid import uuid4
 import django
 django.setup()
 # dashboard
-from django.contrib.auth.models import User
 from django.utils import timezone
-from main.models import Agent, Derivation, DublinCore, Event, File, FileID, FPCommandOutput, SIP, SIPArrange, Transfer
+from main.models import Agent, Derivation, Directory, DublinCore, Event, File, FileID, FPCommandOutput, SIP, SIPArrange
 
+import archivematicacreateaipmetsdataverse as dataverseaip
 import archivematicaCreateMETSReingest
-from createMETSDataverse import create_dataverse_sip_dmdsec, create_dataverse_tabfile_dmdsec
 from archivematicaCreateMETSMetadataCSV import parseMetadata
 from archivematicaCreateMETSRights import archivematicaGetRights
 from archivematicaCreateMETSRightsDspaceMDRef import archivematicaCreateMETSRightsDspaceMDRef
@@ -60,16 +61,16 @@ sharedVariablesAcrossModules.globalErrorCount = 0
 
 from bagit import Bag, BagError
 
-#Global Variables
+# Global Variables
 
 globalFileGrps = {}
-globalFileGrpsUses = ["original", "submissionDocumentation", "preservation", "service", "access", "license", "text/ocr", "metadata", "derivative"]
+globalFileGrpsUses = ["original", "submissionDocumentation", "preservation", "service", "access", "license", "text/ocr", "metadata"]
 for use in globalFileGrpsUses:
     grp = etree.Element(ns.metsBNS + "fileGrp")
     grp.set("USE", use)
     globalFileGrps[use] = grp
 
-##counters
+# counters
 global amdSecs
 amdSecs = []
 global dmdSecs
@@ -84,19 +85,30 @@ global globalRightsMDCounter
 globalRightsMDCounter = 0
 global globalDigiprovMDCounter
 globalDigiprovMDCounter = 0
-global fileNameToFileID #Used for mapping structMaps included with transfer
+global fileNameToFileID  # Used for mapping structMaps included with transfer
 fileNameToFileID = {}
+global globalStructMapCounter
+globalStructMapCounter = 0
 
 global trimStructMap
 trimStructMap = None
 global trimStructMapObjects
 trimStructMapObjects = None
-#GROUPID="G1" -> GROUPID="Group-%object's UUID%"
-##group of the object and it's related access, license
+# GROUPID="G1" -> GROUPID="Group-%object's UUID%"
+# group of the object and it's related access, license
 
 CSV_METADATA = {}
 
-#move to common
+# move to common
+
+
+logger = get_script_logger("archivematica.mcp.client.createMETS2")
+
+
+FSItem = collections.namedtuple('FSItem', 'type path is_empty')
+FakeDirMdl = collections.namedtuple('FakeDirMdl', 'uuid')
+
+
 def newChild(parent, tag, text=None, tailText=None, sets=[]):
     # TODO convert sets to a dict, and use **dict
     child = etree.SubElement(parent, tag)
@@ -107,7 +119,6 @@ def newChild(parent, tag, text=None, tailText=None, sets=[]):
         key, value = set
         child.set(key, value)
     return child
-
 
 
 SIPMetadataAppliesToType = '3e48343d-e2d2-4956-aaa3-b54d26eb9761'
@@ -154,6 +165,64 @@ def getDublinCore(unit, id):
             elem_ns = ns.dctermsBNS
         if txt:
             newChild(ret, elem_ns + term, text=txt)
+    return ret
+
+
+def _get_mdl_identifiers(mdl):
+    """Get identifiers of a model as a type-value 2-tuple."""
+    if isinstance(mdl, FakeDirMdl):
+        return []
+    return [(idfr.type, idfr.value) for idfr in mdl.identifiers.all()]
+
+
+def _add_identifier(object_elem, identifier, bns=ns.premisBNS):
+    """Add an identifier to a <premis:object>."""
+    idfr_type, idfr_val = identifier
+    objectIdentifier = etree.SubElement(
+        object_elem,
+        bns + "objectIdentifier")
+    etree.SubElement(
+        objectIdentifier,
+        bns + "objectIdentifierType").text = idfr_type
+    etree.SubElement(
+        objectIdentifier,
+        bns + "objectIdentifierValue").text = idfr_val
+    return object_elem
+
+
+def getDirDmdSec(dir_mdl, relativeDirectoryPath):
+    """Return an lxml ``Element`` representing a <mets:dmdSec> for a directory.
+    It describes the directory as a PREMIS:OBJECT of type
+    premis:intellectualEntity and lists the directory's original name (i.e.,
+    relative path within the transfer) as well as the UUID assigned to it during
+    transfer (or arrange), and any other identifiers, as
+    premis:objectIdentifiers. Cf. https://projects.artefactual.com/issues/11192.
+    """
+    dir_uuid = dir_mdl.uuid
+    ret = etree.Element(ns.metsBNS + "dmdSec")
+    mdWrap = etree.SubElement(ret, ns.metsBNS + "mdWrap")
+    mdWrap.set("MDTYPE", "PREMIS:OBJECT")
+    xmlData = etree.SubElement(mdWrap, ns.metsBNS + "xmlData")
+    object_elem = etree.SubElement(xmlData, ns.premisV3BNS + "object",
+                                   nsmap={'premis': ns.premisV3NS})
+    object_elem.set(ns.xsiBNS + "type", "premis:intellectualEntity")
+    object_elem.set(
+        ns.xsiBNS + "schemaLocation",
+        ns.premisV3NS + " http://www.loc.gov/standards/premis/v3/premis.xsd")
+    object_elem.set("version", "3.0")
+    # Add the directory's UUID and any other identifiers as
+    # <premis:objectIdentifier> children of <premis:object>.
+    for identifier in chain((('UUID', dir_uuid),),
+                            _get_mdl_identifiers(dir_mdl)):
+        object_elem = _add_identifier(
+            object_elem, identifier, bns=ns.premisV3BNS)
+    try:
+        original_name = escape(dir_mdl.originallocation)
+    except AttributeError:  # SIP model won't have originallocation
+        original_name = escape(relativeDirectoryPath)
+    etree.SubElement(
+        object_elem,
+        ns.premisV3BNS + "originalName").text = original_name
     return ret
 
 
@@ -281,6 +350,7 @@ def createDublincoreDMDSecFromDBData(unit_type, unit_uuid, baseDirectoryPath):
     xmlData.append(dc)
     return (dmdSec, ID)
 
+
 def createDSpaceDMDSec(label, dspace_mets_path, directoryPathSTR):
     """
     Parse DSpace METS file and create a dmdSecs from the info.
@@ -351,6 +421,7 @@ def createTechMD(fileUUID):
     xmlData.append(premis_object)
     return ret
 
+
 def create_premis_object(fileUUID):
     """
     Create a PREMIS:OBJECT for fileUUID.
@@ -363,13 +434,14 @@ def create_premis_object(fileUUID):
     f = File.objects.get(uuid=fileUUID)
     # PREMIS:OBJECT
     object_elem = etree.Element(ns.premisBNS + "object", nsmap={'premis': ns.premisNS})
-    object_elem.set(ns.xsiBNS+"type", "premis:file")
-    object_elem.set(ns.xsiBNS+"schemaLocation", ns.premisNS + " http://www.loc.gov/standards/premis/v2/premis-v2-2.xsd")
+    object_elem.set(ns.xsiBNS + "type", "premis:file")
+    object_elem.set(ns.xsiBNS + "schemaLocation", ns.premisNS + " http://www.loc.gov/standards/premis/v2/premis-v2-2.xsd")
     object_elem.set("version", "2.2")
 
-    objectIdentifier = etree.SubElement(object_elem, ns.premisBNS + "objectIdentifier")
-    etree.SubElement(objectIdentifier, ns.premisBNS + "objectIdentifierType").text = "UUID"
-    etree.SubElement(objectIdentifier, ns.premisBNS + "objectIdentifierValue").text = fileUUID
+    # Add the UUID and any additional file identifiers, e.g., PIDs or
+    # PURLs/URIs, to the XML.
+    for identifier in chain((('UUID', fileUUID),), _get_mdl_identifiers(f)):
+        object_elem = _add_identifier(object_elem, identifier)
 
     objectCharacteristics = etree.SubElement(object_elem, ns.premisBNS + "objectCharacteristics")
     etree.SubElement(objectCharacteristics, ns.premisBNS + "compositionLevel").text = "0"
@@ -382,6 +454,11 @@ def create_premis_object(fileUUID):
 
     for elem in create_premis_object_formats(fileUUID):
         objectCharacteristics.append(elem)
+
+    creatingApplication = etree.Element(ns.premisBNS + "creatingApplication")
+    etree.SubElement(creatingApplication, ns.premisBNS + "dateCreatedByApplication").text = f.modificationtime.strftime("%Y-%m-%d")
+    objectCharacteristics.append(creatingApplication)
+
     for elem in create_premis_object_characteristics_extensions(fileUUID):
         objectCharacteristics.append(elem)
 
@@ -431,6 +508,7 @@ def create_premis_object_characteristics_extensions(fileUUID):
 
     return elements
 
+
 def create_premis_object_derivations(fileUUID):
     elements = []
     # Derivations
@@ -468,6 +546,7 @@ def create_premis_object_derivations(fileUUID):
 
     return elements
 
+
 def createDigiprovMD(fileUUID):
     """
     Create digiprovMD for PREMIS Events and linking Agents.
@@ -497,6 +576,7 @@ def createDigiprovMD(fileUUID):
 
     return ret
 
+
 def createEvent(event_record):
     """ Returns a PREMIS Event. """
     event = etree.Element(ns.premisBNS + "event", nsmap={'premis': ns.premisNS})
@@ -523,10 +603,11 @@ def createEvent(event_record):
         etree.SubElement(linkingAgentIdentifier, ns.premisBNS + "linkingAgentIdentifierValue").text = agent.identifiervalue
     return event
 
+
 def createAgent(agent_record):
     """ Creates a PREMIS Agent as a SubElement of digiprovMD. """
     agent = etree.Element(ns.premisBNS + "agent", nsmap={'premis': ns.premisNS})
-    agent.set(ns.xsiBNS+"schemaLocation", ns.premisNS + " http://www.loc.gov/standards/premis/v2/premis-v2-2.xsd")
+    agent.set(ns.xsiBNS + "schemaLocation", ns.premisNS + " http://www.loc.gov/standards/premis/v2/premis-v2-2.xsd")
     agent.set("version", "2.2")
 
     agentIdentifier = etree.SubElement(agent, ns.premisBNS + "agentIdentifier")
@@ -535,6 +616,7 @@ def createAgent(agent_record):
     etree.SubElement(agent, ns.premisBNS + "agentName").text = agent_record.name
     etree.SubElement(agent, ns.premisBNS + "agentType").text = agent_record.agenttype
     return agent
+
 
 def getAMDSec(fileUUID, filePath, use, type, sip_uuid, transferUUID, itemdirectoryPath, typeOfTransfer, baseDirectoryPath):
     """
@@ -562,13 +644,13 @@ def getAMDSec(fileUUID, filePath, use, type, sip_uuid, transferUUID, itemdirecto
     AMD = etree.Element(ns.metsBNS + "amdSec", ID=AMDID)
     ret = (AMD, AMDID)
 
-    #tech MD
+    # tech MD
     AMD.append(createTechMD(fileUUID))
 
     if use == "original":
         metadataAppliesToList = [(fileUUID, FileMetadataAppliesToType), (sip_uuid, SIPMetadataAppliesToType), (transferUUID, TransferMetadataAppliesToType)]
         for a in archivematicaGetRights(metadataAppliesToList, fileUUID):
-            globalRightsMDCounter +=1
+            globalRightsMDCounter += 1
             rightsMD = etree.SubElement(AMD, ns.metsBNS + "rightsMD")
             rightsMD.set("ID", "rightsMD_" + globalRightsMDCounter.__str__())
             mdWrap = newChild(rightsMD, ns.metsBNS + "mdWrap")
@@ -578,7 +660,7 @@ def getAMDSec(fileUUID, filePath, use, type, sip_uuid, transferUUID, itemdirecto
 
         if typeOfTransfer == "Dspace":
             for a in archivematicaCreateMETSRightsDspaceMDRef(fileUUID, filePath, transferUUID, itemdirectoryPath):
-                globalRightsMDCounter +=1
+                globalRightsMDCounter += 1
                 rightsMD = etree.SubElement(AMD, ns.metsBNS + "rightsMD")
                 rightsMD.set("ID", "rightsMD_" + globalRightsMDCounter.__str__())
                 rightsMD.append(a)
@@ -586,13 +668,14 @@ def getAMDSec(fileUUID, filePath, use, type, sip_uuid, transferUUID, itemdirecto
         elif typeOfTransfer == "TRIM":
             digiprovMD = getTrimFileAmdSec(baseDirectoryPath, sip_uuid, fileUUID)
             globalDigiprovMDCounter += 1
-            digiprovMD.set("ID", "digiprovMD_"+ globalDigiprovMDCounter.__str__())
+            digiprovMD.set("ID", "digiprovMD_" + globalDigiprovMDCounter.__str__())
             AMD.append(digiprovMD)
 
     for a in createDigiprovMD(fileUUID):
         AMD.append(a)
 
     return ret
+
 
 def getIncludedStructMap(baseDirectoryPath):
     global fileNameToFileID
@@ -612,8 +695,8 @@ def getIncludedStructMap(baseDirectoryPath):
             continue
         if os.path.isfile(structMapXmlPath):
             tree = etree.parse(structMapXmlPath)
-            root = tree.getroot() #TDOD - not root to return, but sub element structMap
-            #print etree.tostring(root)
+            root = tree.getroot()  # TDOD - not root to return, but sub element structMap
+            # print etree.tostring(root)
             structMap = root.find(ns.metsBNS + "structMap")
             id = structMap.get("ID")
             if not id:
@@ -622,24 +705,28 @@ def getIncludedStructMap(baseDirectoryPath):
             for item in structMap.findall(".//" + ns.metsBNS + "fptr"):
                 fileName = item.get("FILEID")
                 if fileName in fileNameToFileID:
-                    #print fileName, " -> ", fileNameToFileID[fileName]
+                    # print fileName, " -> ", fileNameToFileID[fileName]
                     item.set("FILEID", fileNameToFileID[fileName])
                 else:
                     print("error: no fileUUID for ", fileName, file=sys.stderr)
                     sharedVariablesAcrossModules.globalErrorCount += 1
-    for fileName, fileID in  fileNameToFileID.items():
-        #locate file based on key
+    for fileName, fileID in fileNameToFileID.items():
+        # locate file based on key
         continue
         print(fileName)
-    if trimStructMap != None:
+    if trimStructMap is not None:
         ret.append(trimStructMap)
     return ret
 
-#DMDID="dmdSec_01" for an object goes in here
-#<file ID="file1-UUID" GROUPID="G1" DMDID="dmdSec_02" ADMID="amdSec_01">
-def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName, fileGroupIdentifier, fileGroupType, includeAmdSec=True):
-    """
-    Creates fileSec and structMap entries for files on disk recursively.
+# DMDID="dmdSec_01" for an object goes in here
+# <file ID="file1-UUID" GROUPID="G1" DMDID="dmdSec_02" ADMID="amdSec_01">
+
+
+def createFileSec(directoryPath, parentDiv, baseDirectoryPath,
+                  baseDirectoryName, fileGroupIdentifier, fileGroupType,
+                  directories, includeAmdSec=True):
+
+    """Creates fileSec and structMap entries for files on disk recursively.
 
     :param directoryPath: Path to recursively traverse and create METS entries for
     :param parentDiv: structMap div to attach created children to
@@ -667,16 +754,44 @@ def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName
         print(directoryPath, "doesn't exist", file=sys.stderr)
         return
 
-    structMapDiv = etree.SubElement(parentDiv, ns.metsBNS + 'div', TYPE='Directory', LABEL=os.path.basename(directoryPath))
+    # Create the <mets:div> element for the directory that this file is in.
+    # If this directory has been assigned a UUID during transfer, retrieve that
+    # UUID based on the directory's relative path and document it in its own
+    # <mets:dmdSec> element.
+    directoryName = os.path.basename(directoryPath)
+    relativeDirectoryPath = (
+        '%SIPDirectory%' +
+        os.path.join(directoryPath.replace(baseDirectoryPath, "", 1), ''))
+    dir_mdl = directories.get(
+        relativeDirectoryPath,
+        directories.get(relativeDirectoryPath.rstrip('/')))
+    dir_dmd_id = None
+    if dir_mdl:
+        dirDmdSec = getDirDmdSec(dir_mdl, relativeDirectoryPath)
+        globalDmdSecCounter += 1
+        dmdSecs.append(dirDmdSec)
+        dir_dmd_id = "dmdSec_" + globalDmdSecCounter.__str__()
+        dirDmdSec.set("ID", dir_dmd_id)
+    structMapDiv = etree.SubElement(
+        parentDiv, ns.metsBNS + 'div', TYPE='Directory',
+        LABEL=directoryName)
 
     DMDIDS = createDMDIDsFromCSVMetadata(directoryPath.replace(baseDirectoryPath, "", 1))
-    if DMDIDS:
-        structMapDiv.set("DMDID", DMDIDS)
+    if DMDIDS or dir_dmd_id:
+        if DMDIDS and dir_dmd_id:
+            structMapDiv.set("DMDID", dir_dmd_id + ' ' + DMDIDS)
+        elif DMDIDS:
+            structMapDiv.set("DMDID", DMDIDS)
+        else:
+            structMapDiv.set("DMDID", dir_dmd_id)
 
     for item in directoryContents:
         itemdirectoryPath = os.path.join(directoryPath, item)
         if os.path.isdir(itemdirectoryPath):
-            createFileSec(itemdirectoryPath, structMapDiv, baseDirectoryPath, baseDirectoryName, fileGroupIdentifier, fileGroupType, includeAmdSec)
+            createFileSec(itemdirectoryPath, structMapDiv, baseDirectoryPath,
+                          baseDirectoryName, fileGroupIdentifier,
+                          fileGroupType, directories, includeAmdSec)
+
         elif os.path.isfile(itemdirectoryPath):
             # Setup variables for creating file metadata
             DMDIDS = ""
@@ -690,9 +805,9 @@ def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName
             try:
                 f = File.objects.get(**kwargs)
             except File.DoesNotExist:
-                print("No uuid for file: \"", directoryPathSTR, "\"", file=sys.stderr)
+                print('No uuid for file: "', directoryPathSTR, '"',
+                      file=sys.stderr)
                 sharedVariablesAcrossModules.globalErrorCount += 1
-                continue
 
             use = f.filegrpuse
             label = f.label
@@ -774,15 +889,18 @@ def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName
                     "filegrpuse": "original",
                     "originallocation__startswith": os.path.dirname(f.originallocation)
                 }
-                try:
-                    original_file = File.objects.get(**kwargs)
+                original_file = File.objects.filter(**kwargs).first()
+                if original_file is not None:
                     GROUPID = 'Group-' + original_file.uuid
-                except (File.DoesNotExist, File.MultipleObjectsReturned):
-                    pass
 
-            elif use in ("preservation", "text/ocr", "derivative"):
-                # Derived files (by AM or external) should be in the original file's group
-                d = Derivation.objects.get(derived_file_id=f.uuid)
+            elif use in ("preservation", "text/ocr"):
+                # Derived files should be in the original file's group
+                try:
+                    d = Derivation.objects.get(derived_file_id=f.uuid)
+                except Derivation.DoesNotExist:
+                    print('Fatal error: unable to locate a Derivation object'
+                          ' where the derived file is {}'.format(f.uuid))
+                    raise
                 GROUPID = "Group-" + d.source_file_id
 
             elif use == "service":
@@ -822,13 +940,6 @@ def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName
                     else:
                         dspaceMetsDMDID = ids
 
-            # If it's a .tab file, check if there's a Dataverse METS with additional metadata
-            if f.originallocation.endswith('.tab'):
-                dv_metadata = create_dataverse_tabfile_dmdsec(baseDirectoryPath, os.path.basename(f.originallocation))
-                dmdSecs.extend(dv_metadata)
-                ids = ' '.join([ds.get('ID') for ds in dv_metadata])
-                fileDiv.attrib['DMDID'] = fileDiv.attrib.get('DMDID', '') + ' ' + ids
-
             if GROUPID == "":
                 sharedVariablesAcrossModules.globalErrorCount += 1
                 print("No groupID for file: \"", directoryPathSTR, "\"", file=sys.stderr)
@@ -840,7 +951,7 @@ def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName
                 file_elem = etree.SubElement(globalFileGrps[use], ns.metsBNS + "file", ID=fileId, GROUPID=GROUPID)
                 if use == "original":
                     filesInThisDirectory.append(file_elem)
-                #<Flocat xlink:href="objects/file1-UUID" locType="other" otherLocType="system"/>
+                # <Flocat xlink:href="objects/file1-UUID" locType="other" otherLocType="system"/>
                 newChild(file_elem, ns.metsBNS + "FLocat", sets=[(ns.xlinkBNS + "href", directoryPathSTR), ("LOCTYPE", "OTHER"), ("OTHERLOCTYPE", "SYSTEM")])
                 if includeAmdSec:
                     AMD, ADMID = getAMDSec(f.uuid, directoryPathSTR, use, fileGroupType, fileGroupIdentifier, f.transfer_id, itemdirectoryPath, typeOfTransfer, baseDirectoryPath)
@@ -852,6 +963,7 @@ def createFileSec(directoryPath, parentDiv, baseDirectoryPath, baseDirectoryName
             file_elem.set("DMDID", dspaceMetsDMDID)
 
     return structMapDiv
+
 
 def build_arranged_structmap(original_structmap, sip_uuid):
     """
@@ -913,6 +1025,7 @@ def build_arranged_structmap(original_structmap, sip_uuid):
 
     return structmap
 
+
 def find_source_metadata(path):
     """
     Returns lists of all metadata to be referenced in the final document.
@@ -934,12 +1047,14 @@ def find_source_metadata(path):
 
     return transfer, source
 
+
 def find_bag_metadata(bag_logs_path):
     try:
         return Bag(bag_logs_path).info
     except BagError:
         print("Unable to locate or parse bag metadata at: {}".format(bag_logs_path), file=sys.stderr)
         return {}
+
 
 def create_object_metadata(struct_map, baseDirectoryPath):
     transfer_metadata_path = os.path.join(baseDirectoryPath, "objects/metadata/transfers")
@@ -1009,7 +1124,7 @@ def write_mets(tree, filename):
     :param ElementTree tree: METS ElementTree
     :param str filename: Filename to write the METS to
     """
-    tree.write(filename, pretty_print=True, xml_declaration=True)
+    tree.write(filename, pretty_print=True, xml_declaration=True, encoding='utf-8')
 
     import cgi
     validate_filename = filename + ".validatorTester.html"
@@ -1025,22 +1140,128 @@ def write_mets(tree, filename):
     <br/>
   </form>
 </body>
-</html>""" % (cgi.escape(etree.tostring(tree, pretty_print=True, xml_declaration=True)))
+</html>""" % (cgi.escape(etree.tostring(tree, pretty_print=True, xml_declaration=True, encoding='utf-8')))
     with open(validate_filename, 'w') as f:
         f.write(fileContents)
 
+
+def get_paths_as_fsitems(baseDirectoryPath, objectsDirectoryPath):
+    """Get all paths in the SIP as ``FSItem`` instances before deleting any
+    empty directories. These filesystem items are crucially ordered so that
+    directories always precede the paths of the items they contain.
+    :param string baseDirectoryPath: path to the AIP with a trailing slash
+    :param string objectsDirectoryPath: path to the AIP's object directory
+    :returns: list of ``FSItem`` instances representing paths
+    """
+    all_fsitems = []
+    for root, dirs, files in os.walk(objectsDirectoryPath):
+        root = root.replace(baseDirectoryPath, '', 1)
+        if files or dirs:
+            all_fsitems.append(FSItem('dir', root, is_empty=False))
+        else:
+            all_fsitems.append(FSItem('dir', root, is_empty=True))
+        for file_ in files:
+            all_fsitems.append(
+                FSItem('file', os.path.join(root, file_), is_empty=False))
+    return all_fsitems
+
+
+def get_normative_structmap(baseDirectoryPath, objectsDirectoryPath,
+                            directories):
+    """Get a normative structMap representing the paths within a SIP.
+    :param string baseDirectoryPath: path to the AIP with a trailing slash
+    :param string objectsDirectoryPath: path to the AIP's object directory
+    :param dict directories: maps directory model instance ``currentlocation``
+    :returns: etree Element representing structMap XML
+    """
+    normativeStructMap = etree.Element(
+        ns.metsBNS + 'structMap',
+        TYPE='logical',
+        ID='structMap_{}'.format(globalStructMapCounter),
+        LABEL='Normative Directory Structure')
+    normativeStructMapDiv = etree.SubElement(
+        normativeStructMap,
+        ns.metsBNS + 'div',
+        TYPE='Directory',
+        LABEL=os.path.basename(baseDirectoryPath.rstrip('/')))
+    all_fsitems = get_paths_as_fsitems(baseDirectoryPath, objectsDirectoryPath)
+    add_normative_structmap_div(all_fsitems, normativeStructMapDiv, directories)
+    return normativeStructMap
+
+
+def add_normative_structmap_div(all_fsitems, root_el, directories, path_to_el=None):
+    """Document all of the file/dir paths in ``all_fsitems`` in the
+    lxml._Element instance ``root_el``. This constructs the <mets:div> element
+    tree under the TYPE "logical" structMap with LABEL "Normative Directory
+    Structure". Said structural map documents all files and directories,
+    including empty directories, which latter are not documented in the
+    physical structMap.
+    :param list all_fsitems: contains ``FSItem`` instances crucially ordered so
+        that parent directories always precede their children.
+    :param lxml._Element root_el: root element for documenting the directory
+        structure.
+    :param dict directories: maps directory model instance ``currentlocation``
+        values to directories for any and all directory model instances
+        associated to the current SIP.
+    :param dict path_to_el: maps paths from ``all_fsitems`` to the lxml elements
+        that document them.
+    :returns: None.
+    """
+    global globalDmdSecCounter
+    global dmdSecs
+    path_to_el = path_to_el or {'': root_el}
+    for fsitem in all_fsitems:
+        parent_path = os.path.dirname(fsitem.path)
+        basename = os.path.basename(fsitem.path)
+        try:
+            parent_el = path_to_el[parent_path]
+        except KeyError:
+            logger.info('Unable to find parent path {} of item {} in path_to_el\n{}'.format(
+                parent_path, fsitem.path, pprint.pformat(path_to_el)))
+            raise
+        el = etree.SubElement(
+            parent_el,
+            ns.metsBNS + 'div',
+            TYPE={'dir': 'Directory'}.get(fsitem.type, 'Item'),
+            LABEL=basename)
+        if fsitem.is_empty:  # Create dmdSec for empty dirs
+            if fsitem.path.startswith('objects/metadata/transfers/'):
+                continue
+            fsitem_path = '%SIPDirectory%' + fsitem.path
+            dir_mdl = directories.get(
+                fsitem_path, directories.get(
+                    fsitem_path.rstrip('/'), FakeDirMdl(uuid=str(uuid4()))))
+            dirDmdSec = getDirDmdSec(dir_mdl, fsitem_path)
+            globalDmdSecCounter += 1
+            dmdSecs.append(dirDmdSec)
+            dir_dmd_id = 'dmdSec_' + str(globalDmdSecCounter)
+            dirDmdSec.set('ID', dir_dmd_id)
+            el.set('DMDID', dir_dmd_id)
+        path_to_el[fsitem.path] = el
+
+
 if __name__ == '__main__':
-    logger = get_script_logger("archivematica.mcp.client.createMETS2")
 
     from optparse import OptionParser
     parser = OptionParser()
-    parser.add_option("--sipType", action="store", dest="sip_type", default="SIP")
-    parser.add_option("-s",  "--baseDirectoryPath", action="store", dest="baseDirectoryPath", default="")
-    parser.add_option("-b",  "--baseDirectoryPathString", action="store", dest="baseDirectoryPathString", default="SIPDirectory") #transferDirectory/
-    parser.add_option("-f",  "--fileGroupIdentifier", action="store", dest="fileGroupIdentifier", default="") #transferUUID/sipUUID
-    parser.add_option("-t",  "--fileGroupType", action="store", dest="fileGroupType", default="sipUUID") #
-    parser.add_option("-x",  "--xmlFile", action="store", dest="xmlFile", default="")
-    parser.add_option("-a",  "--amdSec", action="store_true", dest="amdSec", default=False)
+    parser.add_option("--sipType", action="store", dest="sip_type",
+                      default="SIP")
+    parser.add_option("-s", "--baseDirectoryPath", action="store",
+                      dest="baseDirectoryPath", default="")
+    # transferDirectory/
+    parser.add_option("-b", "--baseDirectoryPathString", action="store",
+                      dest="baseDirectoryPathString", default="SIPDirectory")
+    # transferUUID/sipUUID
+    parser.add_option("-f", "--fileGroupIdentifier", action="store",
+                      dest="fileGroupIdentifier", default="")
+    parser.add_option("-t", "--fileGroupType", action="store",
+                      dest="fileGroupType", default="sipUUID")
+    parser.add_option("-x", "--xmlFile", action="store", dest="xmlFile",
+                      default="")
+    parser.add_option("-a", "--amdSec", action="store_true", dest="amdSec",
+                      default=False)
+    parser.add_option("-n", "--createNormativeStructmap", action="store_true",
+                      dest="createNormativeStructmap", default=False)
     (opts, args) = parser.parse_args()
 
     SIP_TYPE = opts.sip_type
@@ -1050,14 +1271,39 @@ if __name__ == '__main__':
     fileGroupIdentifier = opts.fileGroupIdentifier
     fileGroupType = opts.fileGroupType
     includeAmdSec = opts.amdSec
+    createNormativeStructmap = opts.createNormativeStructmap
+    keepNormativeStructmap = createNormativeStructmap
+
+    # For the Dataverse (DV) proof of concept, if the transfer type is
+    # 'Dataverse' then fork this work to the Dataverse POC METS2 script.
+    # The DV work in that METS2 file needs to be merge into this one, or,
+    # this one refactored with transfer-type specific changes then incorporated
+    # in a more compartmentalised way.
+    for f in File.objects.filter(sip=fileGroupIdentifier):
+        if f.transfer:
+            if f.transfer.type == 'Dataverse':
+                logger.info("Transfer type: %s. Creating %s METS",
+                             f.transfer.type, f.transfer.type)
+                dataverseaip.write_dv_mets(
+                  sip_type=SIP_TYPE,
+                  baseDirectoryPath=baseDirectoryPath,
+                  baseDirectoryPathString=baseDirectoryPathString,
+                  fileGroupIdentifier=fileGroupIdentifier,
+                  fileGroupType=fileGroupType,
+                  xmlFile=XMLFile,
+                  amdSec=includeAmdSec,
+                )
+                sys.exit(0)
 
     # If reingesting, do not create a new METS, just modify existing one
     if 'REIN' in SIP_TYPE:
         print('Updating METS during reingest')
         # fileGroupIdentifier is SIPUUID, baseDirectoryPath is SIP dir,
+        # don't keep existing normative structmap if creating one
         root = archivematicaCreateMETSReingest.update_mets(
             baseDirectoryPath,
             fileGroupIdentifier,
+            keep_normative_structmap=keepNormativeStructmap
         )
         tree = etree.ElementTree(root)
         write_mets(tree, XMLFile)
@@ -1068,6 +1314,34 @@ if __name__ == '__main__':
 
     baseDirectoryPath = os.path.join(baseDirectoryPath, '')
     objectsDirectoryPath = os.path.join(baseDirectoryPath, 'objects')
+    objectsMetadataDirectoryPath = os.path.join(objectsDirectoryPath, 'metadata')
+
+    # Fetch any ``Directory`` objects in the database that are contained within
+    # this SIP and return them as a dict from relative paths to UUIDs. (See
+    # createSIPfromTransferObjects.py for the association of ``Directory``
+    # objects to a ``SIP``.
+    directories = {
+        d.currentlocation.rstrip('/'): d for d in
+        Directory.objects.filter(sip_id=fileGroupIdentifier).all()}
+
+    globalStructMapCounter += 1
+    structMap = etree.Element(
+        ns.metsBNS + "structMap", TYPE='physical',
+        ID='structMap_{}'.format(globalStructMapCounter),
+        LABEL="Archivematica default")
+    sip_dir_name = os.path.basename(baseDirectoryPath.rstrip('/'))
+    structMapDiv = etree.SubElement(
+        structMap, ns.metsBNS + 'div', TYPE="Directory",
+        LABEL=sip_dir_name)
+
+    if createNormativeStructmap:
+        # Create the normative structmap.
+        globalStructMapCounter += 1
+        normativeStructMap = get_normative_structmap(
+            baseDirectoryPath, objectsDirectoryPath, directories)
+    else:
+        print("Skipping creation of normative structmap")
+        normativeStructMap = None
 
     # Delete empty directories, see #8427
     for root, dirs, files in os.walk(baseDirectoryPath, topdown=False):
@@ -1077,9 +1351,21 @@ if __name__ == '__main__':
         except OSError:
             pass
 
-    structMap = etree.Element(ns.metsBNS + "structMap", TYPE='physical', ID='structMap_1', LABEL="Archivematica default")
-    structMapDiv = etree.SubElement(structMap, ns.metsBNS + 'div', TYPE="Directory", LABEL=os.path.basename(baseDirectoryPath.rstrip('/')))
-    structMapDivObjects = createFileSec(objectsDirectoryPath, structMapDiv, baseDirectoryPath, baseDirectoryPathString, fileGroupIdentifier, fileGroupType, includeAmdSec)
+    # Get the <dmdSec> for the entire AIP; it is associated to the root
+    # <mets:div> in the physical structMap.
+    sip_mdl = SIP.objects.filter(uuid=fileGroupIdentifier).first()
+    if sip_mdl:
+        aipDmdSec = getDirDmdSec(sip_mdl, sip_dir_name)
+        globalDmdSecCounter += 1
+        dmdSecs.append(aipDmdSec)
+        aip_dmd_id = "dmdSec_" + str(globalDmdSecCounter)
+        aipDmdSec.set("ID", aip_dmd_id)
+        structMapDiv.set("DMDID", aip_dmd_id)
+
+    structMapDivObjects = createFileSec(
+        objectsDirectoryPath, structMapDiv, baseDirectoryPath,
+        baseDirectoryPathString, fileGroupIdentifier, fileGroupType,
+        directories, includeAmdSec)
 
     el = create_object_metadata(structMapDivObjects, baseDirectoryPath)
     if el:
@@ -1087,10 +1373,12 @@ if __name__ == '__main__':
 
     # In an AIC, the metadata dir is not inside the objects dir
     metadataDirectoryPath = os.path.join(baseDirectoryPath, 'metadata')
-    createFileSec(metadataDirectoryPath, structMapDiv, baseDirectoryPath, baseDirectoryPathString, fileGroupIdentifier, fileGroupType, includeAmdSec)
+    createFileSec(metadataDirectoryPath, structMapDiv, baseDirectoryPath,
+                  baseDirectoryPathString, fileGroupIdentifier, fileGroupType,
+                  directories, includeAmdSec)
 
     fileSec = etree.Element(ns.metsBNS + "fileSec")
-    for group in globalFileGrpsUses: #globalFileGrps.itervalues():
+    for group in globalFileGrpsUses:  # globalFileGrps.itervalues():
         grp = globalFileGrps[group]
         if len(grp) > 0:
             fileSec.append(grp)
@@ -1101,13 +1389,14 @@ if __name__ == '__main__':
         'xlink': ns.xlinkNS,
     }
     root = etree.Element(ns.metsBNS + "mets",
-        nsmap = rootNSMap,
-        attrib = { "{" + ns.xsiNS + "}schemaLocation" : "http://www.loc.gov/METS/ http://www.loc.gov/standards/mets/version18/mets.xsd" },
-    )
-    etree.SubElement(root, ns.metsBNS + "metsHdr").set("CREATEDATE", timezone.now().strftime("%Y-%m-%dT%H:%M:%S"))
+                         nsmap=rootNSMap,
+                         attrib={"{" + ns.xsiNS + "}schemaLocation": "http://www.loc.gov/METS/ http://www.loc.gov/standards/mets/version111/mets.xsd"},
+                         )
+    etree.SubElement(root, ns.metsBNS + "metsHdr").set(
+        "CREATEDATE", timezone.now().strftime("%Y-%m-%dT%H:%M:%S"))
 
     dc = createDublincoreDMDSecFromDBData(SIPMetadataAppliesToType, fileGroupIdentifier, baseDirectoryPath)
-    if dc != None:
+    if dc is not None:
         (dmdSec, ID) = dc
         if structMapDivObjects is not None:
             structMapDivObjects.set("DMDID", ID)
@@ -1116,13 +1405,6 @@ if __name__ == '__main__':
             # Attach the DC metadata to the top level SIP div
             # See #9822 for details
             structMapDiv.set('DMDID', ID)
-        root.append(dmdSec)
-    # Check for external (Dataverse) SIP-level dmdSecs
-    dv = create_dataverse_sip_dmdsec(baseDirectoryPath)
-    for dmdSec in dv:
-        dmdid = dmdSec.attrib['ID']
-        dmdids = structMapDivObjects.get("DMDID", '') + ' ' + dmdid
-        structMapDivObjects.set("DMDID", dmdids)
         root.append(dmdSec)
 
     for dmdSec in dmdSecs:
@@ -1133,6 +1415,9 @@ if __name__ == '__main__':
 
     root.append(fileSec)
     root.append(structMap)
+    if normativeStructMap is not None:
+        root.append(normativeStructMap)
+
     for structMapIncl in getIncludedStructMap(baseDirectoryPath):
         root.append(structMapIncl)
